Further cleanups, use .tar.gz downloads instead of .zip
